First of all, the opened sessions will now be closed. This is necessary to not open session after session without closing it again.
Additionally, the state management of orders and baskets was extended. An order can just be created from a Basket which at least has one BasketItem. If the last BasketItem is removed the Order will be deleted.

An orders basket can be edited as long as the status is pending and the payment process didn't start. When the order is no longer pending a new basket can be created but a order for this basket can just be placed when previous order is finished or has expired (takes 30 min).